% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_vanilla_prediction}
\alias{create_vanilla_prediction}
\title{Create vanilla prediction}
\usage{
create_vanilla_prediction(train, id_vars)
}
\arguments{
\item{pred}{}

\item{real}{}
}
\value{
None
}
\description{
Creates a prediction as the same last 28 days
}
