% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dy_plot}
\alias{dy_plot}
\title{Dygraph plot}
\usage{
dy_plot(prod_ts, ids, rolling_period = 1)
}
\arguments{
\item{prod_ts}{product time series object}

\item{ids}{ids to be plotted}
}
\value{
None
}
\description{
Creates a TS dygraph plot for the products in ids
}
